refactor: parameter page
  - split parameters into sections
  - errors now more precise
