+++ src/renderer/pages/parameters/components/components/FixedWidthFlexBox.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const FixedWidthFlexBox = styled.div`
+  display: flex;
+  justify-content: space-between;
+  white-space: nowrap;
+  gap: 5px;
+  max-width: 600px;
+`;
+
+export default FixedWidthFlexBox;
