+++ src/renderer/pages/parameters/components/components/H3.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+const H3 = styled.h3``;
+
+export default H3;
