+++ src/renderer/pages/parameters/components/components/TitleBox.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+import FixedWidthFlexBox from './FixedWidthFlexBox';
+
+const TitleBox = styled(FixedWidthFlexBox)`
+  margin: 10px auto;
+  border-bottom: 1px solid gray;
+  justify-content: center;
+`;
+
+export default TitleBox;
